Updated to use multiple files and increase OOP nature
