+++ Enemy.py
@@ -1,0 +1,6 @@
+from Entity import Entity
+
+class Enemy(Entity):
+    def __init__(self):
+        super().__init__()
+        pass
