+++ Global.py
@@ -1,0 +1,5 @@
+from Player import Player
+from Enemy import Enemy
+
+player = Player()
+enemy = Enemy()
