+++ Player.py
@@ -1,0 +1,6 @@
+from Entity import Entity
+
+class Player(Entity):
+    def __init__(self):
+        super().__init__()
+        pass
